user profile page design changed and some bugs about creating post are fixed.
